+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Cards/ODSCardView.swift
@@ -45,9 +45,6 @@ public struct ODSCardModel {
         self.description = description
         self.destination = AnyView(destination())
     }
-
-    public static let example = ODSCardModel(title: "Title", image: "", subTitle: "Subtitle", description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt.")
-    public static let exampleMultiline = ODSCardModel(title: "Et galisum fugiat ex omnis officia et iusto eius et animi consequuntur et distinctio magnam id autem exercitationem a sint quibusdam. Aut ipsa autem aut omnis architecto ex ratione provident eum placeat atque qui veniam quos est rerum molestiae.", image: "", subTitle: "Sed quasi illo in quidem consectetur sit consequatur voluptatibus sed adipisci fuga quo voluptatem similique non quidem magni! Et dolores libero ut voluptatem possimus ea minus necessitatibus qui totam culpa vel maxime distinctio id ullam dolor et optio velit", description: "Lorem ipsum dolor sit amet. In similique soluta et corrupti aperiam et ipsum quibusdam aut ducimus beatae aut aliquid perferendis qui sunt quam et dolor maxime. Est laudantium quod sit nihil possimus et quas voluptatem in fugit deserunt sit Quis eligendi aut deserunt voluptatum est quia dolor. Quasi odit et incidunt quis sit quia inventore ut repellat quam hic repellat veritatis est dolorem quia et expedita voluptas. Et internos molestiae ut tempora quod non facere nisi ad doloribus velit ad enim corporis qui molestias dolorum qui galisum velit.")
 }
 
 public struct ODSCardView: View {
@@ -57,86 +54,38 @@ public struct ODSCardView: View {
         self.element = element
     }
 
-    public var body: some View {
-
-        CardInnerView(element: element) {}
-            .cornerRadius(10)
-            .shadow(radius: 8)
-    }
-}
-
-public struct ODSCardImageFirst<ButtonContent>: View where ButtonContent: View {
-    public var element: ODSCardModel
-
-    private var buttonContent: () -> ButtonContent
-
-    public init(element: ODSCardModel, @ViewBuilder buttonContent: @escaping () -> ButtonContent) {
-        self.element = element
-        self.buttonContent = buttonContent
+    var image: Image {
+        if let image = element.image {
+            return Image(image.isEmpty ? "ods_empty" : image,
+                         bundle: image.isEmpty ? Bundle.ods : Bundle.main)
+        } else {
+            return Image("ods_empty", bundle: Bundle.ods)
+        }
     }
 
     public var body: some View {
 
-        CardInnerView(element: element) {
-            buttonContent()
-        }
-        .cornerRadius(10)
-        .shadow(radius: 8)
-    }
-}
-
-struct CardInnerView<ButtonContent>: View where ButtonContent: View {
-    let element: ODSCardModel
-    @ViewBuilder var buttonContent: () -> ButtonContent
-
-    var body: some View {
-
-        VStack(alignment: .leading, spacing: ODSSpacing.none) {
-            if let image = element.image {
-                Image(image.isEmpty ? "ods_empty" : image,
-                      bundle: image.isEmpty ? Bundle.ods : Bundle.main)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-            }
-
-            VStack(alignment: .leading, spacing: ODSSpacing.xs) {
-                Text(element.title)
-                    .odsFont(.bodyBold)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-
-                if let subTitle = element.subTitle, !subTitle.isEmpty {
-                    Text(subTitle)
-                }
-                if let description = element.description, !description.isEmpty {
-                    Text(description).padding(.top)
-                }
-                buttonContent().padding(.top)
-            }
-            .layoutPriority(100)
-            .odsFont(.bodyRegular)
-            .foregroundColor(.primary)
-            .padding()
-        }
-        .background(ODSColor.componentBackground2.color)
+        let model = ODSCardImageFirstModel(title: element.title,
+                                           subtitle: element.subTitle,
+                                           image: image,
+                                           supportingText: element.description)
+        ODSCardImageFirst(model: model)
     }
 }
 
 #if DEBUG
+
+extension ODSCardModel {
+
+    public static let example = ODSCardModel(title: "Title", image: "", subTitle: "Subtitle", description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt.")
+    public static let exampleMultiline = ODSCardModel(title: "Et galisum fugiat ex omnis officia et iusto eius et animi consequuntur et distinctio magnam id autem exercitationem a sint quibusdam. Aut ipsa autem aut omnis architecto ex ratione provident eum placeat atque qui veniam quos est rerum molestiae.", image: "", subTitle: "Sed quasi illo in quidem consectetur sit consequatur voluptatibus sed adipisci fuga quo voluptatem similique non quidem magni! Et dolores libero ut voluptatem possimus ea minus necessitatibus qui totam culpa vel maxime distinctio id ullam dolor et optio velit", description: "Lorem ipsum dolor sit amet. In similique soluta et corrupti aperiam et ipsum quibusdam aut ducimus beatae aut aliquid perferendis qui sunt quam et dolor maxime. Est laudantium quod sit nihil possimus et quas voluptatem in fugit deserunt sit Quis eligendi aut deserunt voluptatum est quia dolor. Quasi odit et incidunt quis sit quia inventore ut repellat quam hic repellat veritatis est dolorem quia et expedita voluptas. Et internos molestiae ut tempora quod non facere nisi ad doloribus velit ad enim corporis qui molestias dolorum qui galisum velit.")
+}
+
 struct CardView_Previews: PreviewProvider {
 
     static var previews: some View {
         ScrollView {
-            VStack(alignment: .leading, spacing: ODSSpacing.none) {
-                ODSCardImageFirst(element: ODSCardModel.example) {
-                    Button {} label: {
-                        ODSGenericButtonContent(topText: "Button", textColor: ODSColor.coreBlack.color)
-                    }
-                    .buttonStyle(ODSFilledButtonStyle())
-                }
-                .padding()
-
-                ODSCardView(element: ODSCardModel.example).padding()
-            }
+            ODSCardView(element: ODSCardModel.example).padding()
         }
     }
 }